Add release-please config for version bumping
Replace the manual `version-bump` script with release-please. Adds the
config + manifest (root `wasp` anchor linking the 7 `@wasp.sh/*` packages,
plus an independent `wasp-app-runner`), version annotations, the workflow,
and a `sync-derived-versions` target for regenerating derived artifacts.
