+++ .release-please-manifest.json
@@ -1,0 +1,4 @@
+{
+  "waspc": "0.24.0",
+  "wasp-app-runner": "0.0.16"
+}
