+++ examples/ask-the-documents/main.wasp.ts
@@ -19,7 +19,7 @@ import { Main } from "./src/pages/MainPage" with { type: "ref" };
 
 export default app({
   name: "askTheDocuments",
-  wasp: { version: "^0.24.0" },
+  wasp: { version: "^0.24.0" }, // x-release-please-version
   title: "PG Vector Example",
   head: ["<link rel='icon' href='/favicon.ico' />"],
   auth: {
