+++ examples/waspleau/main.wasp.ts
@@ -8,7 +8,7 @@ import { measureLoadTime } from "./src/workers/loadTime" with { type: "ref" };
 
 export default app({
   name: "waspleau",
-  wasp: { version: "^0.24.0" },
+  wasp: { version: "^0.24.0" }, // x-release-please-version
   title: "Waspleau",
   head: ["<link rel='icon' href='/favicon.ico' />"],
   server: {
