+++ waspc/README.md
@@ -427,10 +427,22 @@ You can run this workflow:
 
 If you put `[skip ci]` in commit message, that commit will be ignored by Github Actions.
 
-We also wrote a `new-release` script which you can use to help you with creating new release: you need to provide it with new version (`./new-release 0.3.0`) and it will check that everything is all right, create appropriate tag and push it, therefore triggering CI to create new release on Github.
+We also wrote a `new-release` script for manually tagging a release: you provide it with a new version (`./new-release 0.3.0`) and it will check that everything is all right, create the appropriate tag and push it. For normal releases the `vX.Y.Z` tag is now created by release-please instead (see [Versioning (release-please)](#versioning-release-please)); `new-release` is mainly used for test / Release Candidate tags (e.g. `./new-release 0.24.1-rc.1`).
 
 NOTE: If building of your commit is suddenly taking much longer time, it might be connected with cache on Github Actions.
 If it happens just once every so it is probably nothing to worry about. If it happens consistently, we should look into it.
+
+### Versioning (release-please)
+
+The Wasp version is bumped by [release-please](https://github.com/googleapis/release-please), configured via `release-please-config.json` and `.release-please-manifest.json` at the repo root.
+
+- Bumps are derived from **conventional commit** messages (`feat:`, `fix:`, `feat!:` / `BREAKING CHANGE:`), so commits need to follow that convention. There is no automated enforcement yet.
+- The whole repo bumps in one shared version, driven by the single `waspc` component. release-please watches the `release` branch and opens a **release PR** into `release` that bumps `waspc.cabal`, the `package.json` of each `@wasp.sh/*` lib/package (matched with globbed `extra-files`), and the `wasp: { version }` in each `examples/*/main.wasp.ts`. `wasp-app-runner` is versioned independently (its own commits, its own `wasp-app-runner-v*` tag).
+- release-please **cannot** regenerate derived files (every `package-lock.json` and the e2e snapshots), so the release PR's CI stays red until you regenerate them manually (see the release steps below). Each lib/package `package-lock.json` is refreshed from its bumped `package.json` by `./run sync-derived-versions`.
+- Merging the release PR creates the `vX.Y.Z` tag. Because release-please creates that tag with `GITHUB_TOKEN`, it does **not** auto-trigger `release-create.yaml`; you trigger that build pipeline manually (see below).
+- `ChangeLog.md` is **not** generated by release-please (`skip-changelog`); keep curating it by hand.
+
+The old `./run version-bump` script has been removed; you no longer hand-edit versions. `./run sync-derived-versions` regenerates the derived artifacts from the already-bumped sources.
 
 ### Typical Release Process
 
@@ -441,16 +453,18 @@ Do the non-bold steps when necessary (decide for each step depending on the chan
 - 👉 Ensure that you have merged any changes from the `release` branch into `main`. You can see the latest PR at https://github.com/wasp-lang/wasp/pull/release.
 - 👉 Update your local repository state to have all remote changes (`git fetch`) and ensure local `main` is up to date.
 - 👉 Branch out from the latest commit you want to release (most likely latest `main`) into a new RC branch called `rc-<version>` (e.g., `rc-0.24.1`) and do the rest of the steps from there.
-- 👉 The version in `waspc.cabal` should already be correct, but double check and update it if needed.
-  - If you modify `waspc.cabal`: create a PR, wait for approval and all the checks (CI) to pass. Then squash and merge the PR into the RC branch.
+- 👉 The version is bumped automatically by release-please, so you do **not** hand-edit `waspc.cabal` or any other version file. The next version is computed from the conventional commits and shown in release-please's release PR title (see [Versioning (release-please)](#versioning-release-please)).
 - 👉 Create an RC release do some testing and fixing (see [below](#test-releases-eg-release-candidate)). Continue when everything is fine.
-- 👉 Check commits since the latest release, and consider enriching and polishing the `ChangeLog.md` and migration guides.
+- 👉 Check commits since the latest release, and consider enriching and polishing the `ChangeLog.md` and migration guides. (release-please does not touch `ChangeLog.md`; it stays hand-curated.)
 - If this is a major version update, take a versioned "snapshot" of the current docs on the RC branch by running `npm run docusaurus docs:version {version}` in the [web](/web) dir. Check the [README in the `web` dir](https://github.com/wasp-lang/wasp/blob/main/web/README.md) for more details. Commit this change to the RC branch and push it.
-- 👉 Switch to the `release` branch and fast-forward it to the RC branch by running `git merge --ff rc-<version>`.
-- 👉 When you're ready to make the final release, make sure you are on `release` and then run `./new-release 0.x.y`.
-  - This script will do some checks, tag the commit with the new release version, and push the tag.
-- 👉 Wait for CI to finish & succeed for the new tag.
-  - This is triggered automatically on tag push. When it's done, it will create a new draft release.
+- 👉 Switch to the `release` branch and fast-forward it to the RC branch by running `git merge --ff rc-<version>`, then push `release`. This is what release-please watches: it opens (or updates) a **release PR** into `release` that bumps every version file.
+- 👉 Regenerate the derived files that release-please can't, on the release PR branch (`release-please--branches--release`). Its CI stays red until you do this:
+  - `cd waspc && ./run build` (the bumped `waspc.cabal` and `package.json`s are already on the branch).
+  - `./run sync-derived-versions` to rebuild the libs/packages (refreshing their own `package-lock.json`) and refresh each `examples/*/package-lock.json`.
+  - `./run test:waspc:e2e:accept-all` to regenerate the e2e snapshots (needs Docker running; `accept-all` can hide mismatches, so also run the comparison-mode e2e test to confirm only version strings changed).
+  - Commit `examples/*/package-lock.json` + `waspc/e2e-tests/test-outputs/snapshots/**` and push to the PR branch.
+- 👉 Review and merge the release PR. release-please then creates the `vX.Y.Z` tag.
+- 👉 Trigger the release build for the new tag. Because release-please creates the tag with `GITHUB_TOKEN`, `release-create.yaml` is **not** auto-triggered: start it manually (re-push the tag, or run the workflow from the Actions UI). When it's done, it will create a new draft release.
 - 👉 Find the new draft release here: https://github.com/wasp-lang/wasp/releases and edit it with your release notes. This usually means copy-pasting the ChangeLog entries for the released version.
 - 👉 Publish the draft release when ready.
 - 👉 Run `npm dist-tag add @wasp.sh/wasp-cli@<version> latest` for users to get the newest version when they install through `npm`.
